+++ Grid.h
@@ -1,0 +1,7 @@
+// Grid.h
+#pragma once
+
+#include "framework.h"
+
+// Функция для обработки кликов мыши
+void HandleMouseClick(HWND hWnd, int xPos, int yPos);
