+++ Renderer.h
@@ -1,0 +1,10 @@
+// Renderer.h
+#pragma once
+
+#include "framework.h"
+#include <gl/GL.h>
+
+// Функции для работы с OpenGL
+void InitOpenGL(HWND hWnd);
+void CleanupOpenGL(HWND hWnd);
+void DrawGrid();
