+++ framework.h
@@ -4,6 +4,8 @@
 
 #pragma once
 
+#pragma comment(lib, "opengl32.lib")
+
 #include "targetver.h"
 #define WIN32_LEAN_AND_MEAN             // Исключите редко используемые компоненты из заголовков Windows
 // Файлы заголовков Windows
@@ -14,3 +16,5 @@
 #include <malloc.h>
 #include <memory.h>
 #include <tchar.h>
+
+#include "resource.h"
