--- life2.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "resource.h"
